api/requirements.txt
django==4.1.5
Pillow==9.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2
django-cors-headers==3.13
django-rest-passwordreset==1.3
djangorestframework-simplejwt==5.2.2
django-environ==0.9
six==1.16
